add redisson lock for like

+++ src/main/java/com/fzq/xiaopotato/service/impl/LikesServiceImpl.java
@@ -2,6 +2,8 @@ package com.fzq.xiaopotato.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+
+import com.esotericsoftware.minlog.Log;
 import com.fzq.xiaopotato.common.ErrorCode;
 import com.fzq.xiaopotato.common.utils.SocketIOUtils;
 import com.fzq.xiaopotato.exception.BusinessException;
@@ -14,12 +16,17 @@ import com.fzq.xiaopotato.model.vo.UserVO;
 import com.fzq.xiaopotato.service.LikesService;
 import com.fzq.xiaopotato.mapper.LikesMapper;
 import com.fzq.xiaopotato.service.UserService;
+import jakarta.annotation.Resource;
 import jakarta.servlet.http.HttpServletRequest;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.DigestUtils;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.concurrent.TimeUnit;
 
 import static com.fzq.xiaopotato.common.NotificationType.FOLLOW;
 import static com.fzq.xiaopotato.common.NotificationType.LIKE;
@@ -46,6 +53,9 @@ public class LikesServiceImpl extends ServiceImpl<LikesMapper, Likes>
     @Autowired
     private SocketIOUtils socketIOUtils;
 
+    @Autowired
+    private RedissonClient redissonClient;
+
     @Override
     public boolean likeByPostId(IdDTO idDTO, HttpServletRequest request) {
         UserVO user = userService.getCurrentUser(request);
@@ -54,34 +64,57 @@ public class LikesServiceImpl extends ServiceImpl<LikesMapper, Likes>
         }
         long userId = user.getId();
         long postId = idDTO.getId();
+        String lockKey = "user:like:" + userId + ":" + postId;
+        RLock lock = redissonClient.getLock(lockKey);
+        try {
+            while (true) {
+                if (lock.tryLock(0, -1, TimeUnit.MILLISECONDS)) {
+                    Log.info("getLock: " + Thread.currentThread().getId());
 
-        QueryWrapper<UserPost> userPostQuery = new QueryWrapper<>();
-        userPostQuery.eq("post_id", postId);
-        UserPost userPost = userPostMapper.selectOne(userPostQuery);
 
-        if (userPost == null) {
-            throw new BusinessException(ErrorCode.PARAMS_ERROR, "Post creator not found");
-        }
+                    QueryWrapper<UserPost> userPostQuery = new QueryWrapper<>();
+                    userPostQuery.eq("post_id", postId);
+                    UserPost userPost = userPostMapper.selectOne(userPostQuery);
 
-        long creatorId = userPost.getUserId();
+                    if (userPost == null) {
+                        throw new BusinessException(ErrorCode.PARAMS_ERROR, "Post creator not found");
+                    }
 
-        boolean isLiked = isPostLikedByUser(userId, postId);
-        if (!isLiked) {
-            Likes likes = new Likes();
-            likes.setUserId(userId);
-            likes.setPostId(postId);
-            likesMapper.insert(likes);
+                    long creatorId = userPost.getUserId();
 
-            if (user.getId() != creatorId) {
-                sendLikeNotification(user, creatorId);
+                    boolean isLiked = isPostLikedByUser(userId, postId);
+                    if (!isLiked) {
+                        Likes likes = new Likes();
+                        likes.setUserId(userId);
+                        likes.setPostId(postId);
+                        likesMapper.insert(likes);
+
+                        if (user.getId() != creatorId) {
+                            sendLikeNotification(user, creatorId);
+                        }
+                        return true;
+                    } else {
+                        QueryWrapper<Likes> queryWrapper = new QueryWrapper<>();
+                        queryWrapper.eq("user_id", userId).eq("post_id", postId);
+                        likesMapper.delete(queryWrapper);
+                        return false;
+                    }
+                }
             }
-            return true;
-        } else {
-            QueryWrapper<Likes> queryWrapper = new QueryWrapper<>();
-            queryWrapper.eq("user_id", userId).eq("post_id", postId);
-            likesMapper.delete(queryWrapper);
-            return false;
+        } catch (InterruptedException e) {
+            log.error(Thread.currentThread().getId() + " get lock failed");
+            throw new BusinessException(ErrorCode.SYSTEM_ERROR, "get lock failed");
+        } finally {
+            // can only release itself lock
+            if (lock.isHeldByCurrentThread()) {
+                lock.unlock();
+                Log.info("release lock: {}", String.valueOf(Thread.currentThread().getId()));
+            }
+
         }
+
+
+
 
     }
